refactor: reorganize file structure and update documentation
- Reorganize project files for improved maintainability
- Add project structure documentation to README.md
